separated log formats.
