+++ src/rikki/patcher/copy/copy_patcher.cpp
@@ -6,6 +6,7 @@
 #include "utils/json_util.hpp"
 
 #include "wv/wv_invoker.hpp"
+#include "wv/wv_log_fmt.hpp"
 
 // ======================== C L A S S ========================
 // ===    CopyPatcher
