+++ src/rikki/patcher/dialogue/dialogue_patcher.cpp
@@ -11,6 +11,7 @@
 #include "utils/filesystem_util.hpp"
 
 #include "wv/wv_invoker.hpp"
+#include "wv/wv_log_fmt.hpp"
 
 // ======================== C L A S S ========================
 // ===    DialoguePatcher
