+++ src/rikki/patcher/ui/ui_patcher.cpp
@@ -8,6 +8,7 @@
 #include "utils/ui_text_util.hpp"
 
 #include "wv/wv_invoker.hpp"
+#include "wv/wv_log_fmt.hpp"
 
 // ======================== C L A S S ========================
 // ===    UIPatcher
