+++ src/rikki/patcher/ui/ui_text_patcher.cpp
@@ -12,6 +12,7 @@
 #include "utils/filesystem_util.hpp"
 
 #include "wv/wv_invoker.hpp"
+#include "wv/wv_log_fmt.hpp"
 
 // ======================== C L A S S ========================
 // ===    UITextPatcher
